refactor(text-input-elements): change how input text is rendered
use the same function as for other text elements
